+++ kitty/state.c
@@ -947,7 +947,7 @@ PYWRAP1(patch_global_colors) {
         else if (PyLong_Check(val)) OPT(name) = PyLong_AsLong(val); \
     } \
 }
-    P(active_border_color); P(inactive_border_color); P(bell_border_color); P(tab_bar_background);
+    P(active_border_color); P(inactive_border_color); P(bell_border_color); P(tab_bar_background); P(tab_bar_margin_color);
     if (configured) {
         P(background); P(url_color);
         P(mark1_background); P(mark1_foreground); P(mark2_background); P(mark2_foreground);