+++ kitty/state.h
@@ -31,7 +31,7 @@ typedef struct {
     unsigned int scrollback_pager_history_size;
     bool scrollback_fill_enlarged_window;
     char_type *select_by_word_characters;
-    color_type url_color, background, foreground, active_border_color, inactive_border_color, bell_border_color, tab_bar_background;
+    color_type url_color, background, foreground, active_border_color, inactive_border_color, bell_border_color, tab_bar_background, tab_bar_margin_color;
     color_type mark1_foreground, mark1_background, mark2_foreground, mark2_background, mark3_foreground, mark3_background;
     monotonic_t repaint_delay, input_delay;
     bool focus_follows_mouse;